Add errors to an :_error key rather than a separate row.
We need to tie the row with the errors up until the point at which we
have assigned a surrogate key to the row, so we know which row to
refer to in the errors.

After doing this assigning of the id, we can demultiplex the errors
into separate rows on the error stream.